requirement.txt
Flask==2.3.0
Flask-Cors==3.1.1
Flask-RESTful==0.3.9
requests==2.31.0

openai==0.27.0
google-cloud==0.34.0